requirements.txt
aiogram==3.0.0
aiohttp==3.9.1
requests==2.31.0
instagrapi==1.16.60
python-dotenv==1.0.0
